+++ src/components/checkboxGroup.js
@@ -98,9 +98,7 @@
       }
     }
 
-    useEffect(() => {
-      B.defineFunction('Refetch', () => refetch());
-    }, [refetch]);
+    B.defineFunction('Refetch', () => refetch());
 
     useEffect(() => {
       if (isDev) {