+++ CHANGELOG.md
@@ -1,3 +1,100 @@
+## [1.56.2](https://github.com/bettyblocks/material-ui-component-set/compare/v1.56.1...v1.56.2) (2020-11-06)
+
+
+### Bug Fixes
+
+* set limit to initial fetches datatable ([#604](https://github.com/bettyblocks/material-ui-component-set/issues/604)) ([9c70195](https://github.com/bettyblocks/material-ui-component-set/commit/9c70195301c80f4154afab39d072a9e97b4595c2))
+
+## [1.56.1](https://github.com/bettyblocks/material-ui-component-set/compare/v1.56.0...v1.56.1) (2020-11-06)
+
+
+### Bug Fixes
+
+* add InteractionScope to grid and list ([2717045](https://github.com/bettyblocks/material-ui-component-set/commit/27170450724bbda3904ff938b0498658e7a76b63))
+
+# [1.56.0](https://github.com/bettyblocks/material-ui-component-set/compare/v1.55.0...v1.56.0) (2020-11-05)
+
+
+### Features
+
+* add visibility option ([9939853](https://github.com/bettyblocks/material-ui-component-set/commit/99398538a3fda8e26716552c0c79e949a1fdc52f))
+
+# [1.55.0](https://github.com/bettyblocks/material-ui-component-set/compare/v1.54.0...v1.55.0) (2020-11-04)
+
+
+### Features
+
+* convert image into media component ([2407644](https://github.com/bettyblocks/material-ui-component-set/commit/2407644052d28067be54cc1fc5f7616933048cba))
+
+# [1.54.0](https://github.com/bettyblocks/material-ui-component-set/compare/v1.53.0...v1.54.0) (2020-11-04)
+
+
+### Features
+
+* change datacontainer to use hooks ([065e8d2](https://github.com/bettyblocks/material-ui-component-set/commit/065e8d2453e280846da0dfdeea65b3c15f735472))
+
+# [1.53.0](https://github.com/bettyblocks/material-ui-component-set/compare/v1.52.0...v1.53.0) (2020-11-03)
+
+
+### Bug Fixes
+
+* autocomplete multiselect labels are selected with default value ([ef58712](https://github.com/bettyblocks/material-ui-component-set/commit/ef587126a56674374b85961ff451f413a200fe05))
+
+
+### Features
+
+* add consistent container spacing options ([f9add32](https://github.com/bettyblocks/material-ui-component-set/commit/f9add32b182455b36dbd5373c51b7563a94f4cbf))
+* add contains and not contains filters to conditional ([54f0352](https://github.com/bettyblocks/material-ui-component-set/commit/54f035210536d15ef2f9f6a366792591cab34d8d))
+
+# [1.52.0](https://github.com/bettyblocks/material-ui-component-set/compare/v1.51.2...v1.52.0) (2020-11-03)
+
+
+### Features
+
+* add contains and not contains filters to conditional ([714f945](https://github.com/bettyblocks/material-ui-component-set/commit/714f94508889c7fb4393f250c2b9c5c043eb11ac))
+
+## [1.51.2](https://github.com/bettyblocks/material-ui-component-set/compare/v1.51.1...v1.51.2) (2020-11-03)
+
+
+### Bug Fixes
+
+* autocomplete multiselect labels are selected with default value ([c179e02](https://github.com/bettyblocks/material-ui-component-set/commit/c179e02e54a7ae400b105a0e1f1b41676d612186))
+
+## [1.51.1](https://github.com/bettyblocks/material-ui-component-set/compare/v1.51.0...v1.51.1) (2020-11-02)
+
+
+### Bug Fixes
+
+* keep dialog mounted in dom for interactions ([b3bb163](https://github.com/bettyblocks/material-ui-component-set/commit/b3bb163d16e8ed96fcfe81dee9411136c3e06302))
+
+# [1.51.0](https://github.com/bettyblocks/material-ui-component-set/compare/v1.50.1...v1.51.0) (2020-10-30)
+
+
+### Features
+
+* add minHeight to box component ([1fb6f2f](https://github.com/bettyblocks/material-ui-component-set/commit/1fb6f2f1cc709b5f87b6d6077d32427eb717f9b4))
+
+## [1.50.1](https://github.com/bettyblocks/material-ui-component-set/compare/v1.50.0...v1.50.1) (2020-10-30)
+
+
+### Bug Fixes
+
+* password visibility toggle ([5aa020f](https://github.com/bettyblocks/material-ui-component-set/commit/5aa020fddeb895c1ff7e4fd2957274da72b2a164))
+
+# [1.50.0](https://github.com/bettyblocks/material-ui-component-set/compare/v1.49.2...v1.50.0) (2020-10-30)
+
+
+### Features
+
+* add style options to checkbox ([def0700](https://github.com/bettyblocks/material-ui-component-set/commit/def070024702beb4b13183253257e78ac34c03b7))
+
+## [1.49.2](https://github.com/bettyblocks/material-ui-component-set/compare/v1.49.1...v1.49.2) (2020-10-30)
+
+
+### Bug Fixes
+
+* set default progress to zero in runtime ([905ede9](https://github.com/bettyblocks/material-ui-component-set/commit/905ede9fef39a8f957598570a885aa7259d944ea))
+
 ## [1.49.1](https://github.com/bettyblocks/material-ui-component-set/compare/v1.49.0...v1.49.1) (2020-10-29)